requirements.txt

fastapi==0.115.0
openpyxl==3.1.5
mysql-connector-python==9.0.0
uvicorn==0.30.6
SQLAlchemy==2.0.35
mysqlclient==2.2.4
requests==2.32.3
pydantic==2.9.2
PyMySQL==1.1.1
